Cleanup. Rename files and links to more intuitive names ("events" instead of "home").
